GCE states/grains/etc
This is a port over of some helpful GCE states and grains we have lying around.
There are states for setting up the repo, the GCE monitoring agent, and the GCE
logging agent.

There are also grains to pull roles/tags from GCE metadata and to pull other
generic GCE metadata from the metadata service.